Add Harvey Nichols matching agent.
This change adds a matching agent for Harvey Nichols. This agent
requires the user's first six and last four digits from the payment
card. In order to have these details available to the matching engine, I
have modified the transaction identifier to pull these details from the
Hermes response.

For testing the system, a HN data provider has been added to the
end-to-end test harness, and the FileAgent type has been implemented.