fix the-works refund export method (#138)
* fix: remove trailing slash from onboard endpoint

this was causing a redirect from /txm/identifiers/ to /txm/identifiers
every time the portal called. the other endpoints do not have trailing
slashes.

* fix: use method 945 to export the-works refunds

fixes refund exports for the-works by using request method 911 when the
amount is positive and 945 when the amount is negative. the request
bodies are the same, so varying the method is sufficient.

both methods seem to accept a positive points value, so i've used
abs(amount) for this.

* test: fix api tests broken by previous commit