feat: preload api-reflector database with endpoints
this works via the following process:

1. manually set up the required endpoint(s) and response(s) in any API
   reflector instance
2. `pg_dump` the database into
   tests/integration/postgres/api_reflector.sql
3. the custom postgres image loads the sql file into
   /usr/local/fixtures/api_reflector.sql
4. on postgres container startup, the init-user-dbs.sh script loads the
   sql file which creates all the tables & inserts the data.

this process should theoretically work for any db that needs static data
pre-populated.

one downside is if the service's schema ever changes, the sql fixture
will need to be updated. this could potentially be done by loading it
into a database and then running migrations.