feat: remove soteria, get secrets from files
this change primarily focuses on reducing or removing reliance on azure
keyvault by instead reading secrets from files. this change means that
soteria is no longer required, so it has been removed. the main reason
for doing this is to allow secrets to be loaded as docker volume mounts
which means testing can be done without any external calls at all.

the iceland and wasabi agents have been left alone (albeit with their
unit tests removed) as they are likely to be removed entirely soon.